+++ src/view/login.js
@@ -10,12 +10,17 @@ export const logInTemplate = () => {
       <input id="emailLogIn" type="email" placeholder="Enter Email" class="inputForm" required>
       <input id="passwordLogIn" type="password" placeholder="Enter Password" class="inputForm" required>
     </form>
-    <button type="submit" class="btnLogIn">Log In</button><br>
+    <button type="button" class="btnLogIn">Log In</button><br>
     <span>New here? <a id="linkRegister" href="#/Register">Register</a></span>
     <div class="divIconG">
       <img src="./img/icons8-logo-de-google.svg" alt="iGoogle" class="iGoogle">
     </div>
     `;
   sectionLogIn.innerHTML = templateLogIn;
+
+  const btnLogIn = sectionLogIn.querySelector('.btnLogIn');
+  btnLogIn.addEventListener('click', () => {
+    console.log('Aqui boton');
+  });
   return sectionLogIn;
 };
